Don't install setuptools separately
This is partly because we shouldn't install outside of the prefix, partly
because the /Library/Python dir is chown:root for people who sudo
easy_installed at some point before.

I don't know if this is the right thing to do or not. So am committing for
discussion and review in a branch.