Add recipe for D version 2.x recipe (dmd2)
This is mostly a copy of dmd.rb, but with some updates for dmd2. Unfortunately
dmd and dmd2 will conflict. The dmd binary looks for dmd.conf explicitly (not
<binname>.conf) and someone that wants dmd and dmd2 installed side by side
probably does not want to share configuration between the two. I'm not sure
how to express the conflicting dependency in a Formula, so a note is added to
caveats.